requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
tensorflow==2.13.0
Werkzeug==2.3.7
setuptools>=65.0.0
Flask
tensorflow
numpy
pillow
